backend/requirements.txt
Flask==2.2.5
python-docx==0.8.11
pdfplumber==0.7.6
pytesseract==0.3.11
Pillow==9.5.0
supabase==0.6.6
python-dotenv==1.0.0
gunicorn==20.1.0   # optional for deployment
